+++ index.d.ts
@@ -10,6 +10,7 @@ declare module '@amaas/amaas-core-sdk-js' {
       function reactivate({ AMId }: { AMId: number }, callback?: Function): Promise<assetManagers.AssetManager> | void
       function checkDomains({ domain }: { domain: string }, callback?: Function): Promise<assetManagers.Domain | null> | void
       function insertDomain({ domain }: { domain: assetManagers.Domain | assetManagers.IDomain }, callback?: Function): Promise<assetManagers.Domain> | void
+      function retrieveEODBooks({ AMId, bookID }: { AMId: number, bookID: string }, callback?: Function): Promise<assetManagers.EODBook | assetManagers.EODBook[]> | void
     }
     namespace Assets {
       function retrieve({ AMId, resourceId }: { AMId: number, resourceId?: string }, callback?: Function): Promise<assets.AssetClassTypes[]> | void
@@ -116,6 +117,31 @@ declare module '@amaas/amaas-core-sdk-js' {
       updatedTime?: string
       version?: number
       constructor(props: IDomain)
+    }
+
+    interface IEODBook {
+      assetManagerId: number
+      utcCloseTime?: string
+      bookId?: string
+      eodBookStatus?: string
+      createdBy?: string
+      updatedBy?: string
+      createdTime?: string
+      updatedTime?: string
+      version?: number
+    }
+
+    class EODBook {
+      assetManagerId: number
+      utcCloseTime?: string
+      bookId?: string
+      eodBookStatus?: string
+      createdBy?: string
+      updatedBy?: string
+      createdTime?: string
+      updatedTime?: string
+      version?: number
+      constructor(props: IEODBook)
     }
   }